+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Improve Python typing throughout
 - Fix height calculation in FPCaptionedView
 - In figpack_slides, support slide editing of titles
+- In figpack_slides, support slide editing of markdown sections
 
 ## [0.2.35] - 2025-10-22